move client auth inside user module
